+++ .github/CHANGELOG.md
@@ -25,9 +25,11 @@
 - Ensure that the variable designating a switch as primary is always initialized.
 
 ## [1.9.0]
+
 - Add CSM 1.6 templates
 - Fix case sensitivity when reading nodes from SHCD
 - Add Storage node support
+- Add config generation support for River NMN nodes
 
 ## [1.8.0]